docs(design-tokens): add elevation tokens page
