+++ src/stories/design-tokens/Elevation.mdx
@@ -1,0 +1,38 @@
+import { Meta } from '@storybook/blocks';
+import { ElevationTokenExample } from './design-token-utils';
+
+<Meta title="Design Tokens/Elevation" />
+
+# Elevation Tokens
+
+Elevation tokens define the shadow and depth of elements.
+
+Each token corresponds to a specific elevation level at a specific height on the z-axis. The higher the elevation level, the higher the element appears to be on the z-axis.
+
+| Token                  | Example                               | Value                                                                                                                 |
+| ---------------------- | ------------------------------------- | --------------------------------------------------------------------------------------------------------------------- |
+| `--forge-elevation-0`  | <ElevationTokenExample radius={0} />  | `0px 0px 0px 0px rgba(0, 0, 0, 0.2), 0px 0px 0px 0px rgba(0, 0, 0, 0.14), 0px 0px 0px 0px rgba(0, 0, 0, 0.12)`        |
+| `--forge-elevation-1`  | <ElevationTokenExample radius={1} />  | `0px 2px 1px -1px rgba(0, 0, 0, 0.2), 0px 1px 1px 0px rgba(0, 0, 0, 0.14), 0px 1px 3px 0px rgba(0, 0, 0, 0.12)`       |
+| `--forge-elevation-2`  | <ElevationTokenExample radius={2} />  | `0px 3px 1px -2px rgba(0, 0, 0, 0.2), 0px 2px 2px 0px rgba(0, 0, 0, 0.14), 0px 1px 5px 0px rgba(0, 0, 0, 0.12)`       |
+| `--forge-elevation-3`  | <ElevationTokenExample radius={3} />  | `0px 3px 3px -2px rgba(0, 0, 0, 0.2), 0px 3px 4px 0px rgba(0, 0, 0, 0.14), 0px 1px 8px 0px rgba(0, 0, 0, 0.12)`       |
+| `--forge-elevation-4`  | <ElevationTokenExample radius={4} />  | `0px 2px 4px -1px rgba(0, 0, 0, 0.2), 0px 4px 5px 0px rgba(0, 0, 0, 0.14), 0px 1px 10px 0px rgba(0, 0, 0, 0.12)`      |
+| `--forge-elevation-5`  | <ElevationTokenExample radius={5} />  | `0px 3px 5px -1px rgba(0, 0, 0, 0.2), 0px 5px 8px 0px rgba(0, 0, 0, 0.14), 0px 1px 14px 0px rgba(0, 0, 0, 0.12)`      |
+| `--forge-elevation-6`  | <ElevationTokenExample radius={6} />  | `0px 3px 5px -1px rgba(0, 0, 0, 0.2), 0px 6px 10px 0px rgba(0, 0, 0, 0.14), 0px 1px 18px 0px rgba(0, 0, 0, 0.12)`     |
+| `--forge-elevation-7`  | <ElevationTokenExample radius={7} />  | `0px 4px 5px -2px rgba(0, 0, 0, 0.2), 0px 7px 10px 1px rgba(0, 0, 0, 0.14), 0px 2px 16px 1px rgba(0, 0, 0, 0.12)`     |
+| `--forge-elevation-8`  | <ElevationTokenExample radius={8} />  | `0px 5px 5px -3px rgba(0, 0, 0, 0.2), 0px 8px 10px 1px rgba(0, 0, 0, 0.14), 0px 3px 14px 2px rgba(0, 0, 0, 0.12);`    |
+| `--forge-elevation-9`  | <ElevationTokenExample radius={9} />  | `0px 5px 6px -3px rgba(0, 0, 0, 0.2), 0px 9px 12px 1px rgba(0, 0, 0, 0.14), 0px 3px 16px 2px rgba(0, 0, 0, 0.12)`     |
+| `--forge-elevation-10` | <ElevationTokenExample radius={10} /> | `0px 6px 6px -3px rgba(0, 0, 0, 0.2), 0px 10px 14px 1px rgba(0, 0, 0, 0.14), 0px 4px 18px 3px rgba(0, 0, 0, 0.12)`    |
+| `--forge-elevation-11` | <ElevationTokenExample radius={11} /> | `0px 6px 7px -4px rgba(0, 0, 0, 0.2), 0px 11px 15px 1px rgba(0, 0, 0, 0.14), 0px 4px 20px 3px rgba(0, 0, 0, 0.12)`    |
+| `--forge-elevation-12` | <ElevationTokenExample radius={12} /> | `0px 7px 8px -4px rgba(0, 0, 0, 0.2), 0px 12px 17px 2px rgba(0, 0, 0, 0.14), 0px 5px 22px 4px rgba(0, 0, 0, 0.12)`    |
+| `--forge-elevation-13` | <ElevationTokenExample radius={13} /> | `0px 7px 8px -4px rgba(0, 0, 0, 0.2), 0px 13px 19px 2px rgba(0, 0, 0, 0.14), 0px 5px 24px 4px rgba(0, 0, 0, 0.12)`    |
+| `--forge-elevation-14` | <ElevationTokenExample radius={14} /> | `0px 7px 9px -4px rgba(0, 0, 0, 0.2), 0px 14px 21px 2px rgba(0, 0, 0, 0.14), 0px 5px 26px 4px rgba(0, 0, 0, 0.12)`    |
+| `--forge-elevation-15` | <ElevationTokenExample radius={15} /> | `0px 8px 9px -5px rgba(0, 0, 0, 0.2), 0px 15px 22px 2px rgba(0, 0, 0, 0.14), 0px 6px 28px 5px rgba(0, 0, 0, 0.12)`    |
+| `--forge-elevation-16` | <ElevationTokenExample radius={16} /> | `0px 8px 10px -5px rgba(0, 0, 0, 0.2), 0px 16px 24px 2px rgba(0, 0, 0, 0.14), 0px 6px 30px 5px rgba(0, 0, 0, 0.12)`   |
+| `--forge-elevation-17` | <ElevationTokenExample radius={17} /> | `0px 8px 11px -5px rgba(0, 0, 0, 0.2), 0px 17px 26px 2px rgba(0, 0, 0, 0.14), 0px 6px 32px 5px rgba(0, 0, 0, 0.12)`   |
+| `--forge-elevation-18` | <ElevationTokenExample radius={18} /> | `0px 9px 11px -5px rgba(0, 0, 0, 0.2), 0px 18px 28px 2px rgba(0, 0, 0, 0.14), 0px 7px 34px 6px rgba(0, 0, 0, 0.12)`   |
+| `--forge-elevation-19` | <ElevationTokenExample radius={19} /> | `0px 9px 12px -6px rgba(0, 0, 0, 0.2), 0px 19px 29px 2px rgba(0, 0, 0, 0.14), 0px 7px 36px 6px rgba(0, 0, 0, 0.12)`   |
+| `--forge-elevation-20` | <ElevationTokenExample radius={20} /> | `0px 10px 13px -6px rgba(0, 0, 0, 0.2), 0px 20px 31px 3px rgba(0, 0, 0, 0.14), 0px 8px 38px 7px rgba(0, 0, 0, 0.12)`  |
+| `--forge-elevation-21` | <ElevationTokenExample radius={21} /> | `0px 10px 13px -6px rgba(0, 0, 0, 0.2), 0px 21px 33px 3px rgba(0, 0, 0, 0.14), 0px 8px 40px 7px rgba(0, 0, 0, 0.12)`  |
+| `--forge-elevation-22` | <ElevationTokenExample radius={22} /> | `0px 10px 14px -6px rgba(0, 0, 0, 0.2), 0px 22px 35px 3px rgba(0, 0, 0, 0.14), 0px 8px 42px 7px rgba(0, 0, 0, 0.12)`  |
+| `--forge-elevation-23` | <ElevationTokenExample radius={23} /> | `0px 11px 14px -7px rgba(0, 0, 0, 0.2), 0px 23px 36px 3px rgba(0, 0, 0, 0.14), 0px 9px 44px 8px rgba(0, 0, 0, 0.12)`  |
+| `--forge-elevation-24` | <ElevationTokenExample radius={24} /> | `0px 11px 15px -7px rgba(0, 0, 0, 0.2), 0px 24px 38px 3px rgba(0, 0, 0, 0.14), 0px 9px 46px 8px rgba(0, 0, 0, 0.12);` |
