+++ src/stories/design-tokens/Shape.mdx
@@ -1,5 +1,5 @@
 import { Meta } from '@storybook/blocks';
-import { ShapeTokenExample } from './shape-utils';
+import { ShapeTokenExample } from './design-token-utils';
 
 <Meta title="Design Tokens/Shape" />
